Fix various compiler warnings
This is a rather big clean-up patch, after Xlint has been set
to 'all', improving type safety (raw types, unchecked casts),
refactoring some bits where necessary, and fixing some stupid
bugs.

+++ bundles/awt-imaging/src/main/java/ch/unifr/diva/dip/awt/imaging/features/Histogram.java
@@ -115,7 +115,7 @@ public class Histogram {
 	public double[] getNormalizedHistogram() {
 		final double[] result = new double[this.counts.length];
 		for (int i = 0; i < this.counts.length; i++) {
-			result[i] = (this.counts[i] / this.totalSamples);
+			result[i] = (this.counts[i] / (double) this.totalSamples);
 		}
 		return result;
 	}
